refactor(components/atom/icon): add import file extensions
