+++ components/atom/icon/demo/settings.js
@@ -1,0 +1,19 @@
+export const BASE_CLASS_DEMO = `DemoAtomIcon`
+export const CLASS_SECTION = `${BASE_CLASS_DEMO}-section`
+
+export const ICONS = {
+  AiOutlineCheck: 'AiOutlineCheck',
+  AiOutlineShop: 'AiOutlineShop',
+  AiOutlineCode: 'AiOutlineCode',
+  AiOutlineCoffee: 'AiOutlineCoffee',
+  AiOutlineLink: 'AiOutlineLink'
+}
+
+export const flexCenteredStyle = {
+  display: 'flex',
+  justifyContent: 'center',
+  flexDirection: 'row',
+  flexWrap: 'nowrap',
+  alignItems: 'center',
+  alignContent: 'center'
+}
