+++ components/atom/icon/src/LazyIcon.js
@@ -1,6 +1,8 @@
 import PropTypes from 'prop-types'
+
 import {useNearScreen} from '@s-ui/react-hooks/lib/useOnScreen'
-import AtomIcon from './Icon'
+
+import AtomIcon from './Icon.js'
 
 export default function LazyIcon({children, ...restOfProps}) {
   const [isIntersecting, outerRef] = useNearScreen()
