+++ components/atom/icon/src/index.js
@@ -1,39 +1,22 @@
 import PropTypes from 'prop-types'
 import cx from 'classnames'
 
-import Icon from './Icon'
-import LazyIcon from './LazyIcon'
+import Icon from './Icon.js'
+import LazyIcon from './LazyIcon.js'
+import {
+  BASE_CLASS,
+  ATOM_ICON_COLORS,
+  ATOM_ICON_SIZES,
+  ATOM_ICON_RENDERS
+} from './settings.js'
 
-const BASE_CLASS = 'sui-AtomIcon'
-export const ATOM_ICON_COLORS = {
-  accent: 'accent',
-  alert: 'alert',
-  currentColor: 'currentColor',
-  error: 'error',
-  primary: 'primary',
-  success: 'success',
-  gray: 'gray'
-}
-export const ATOM_ICON_SIZES = {
-  extraSmall: 'extraSmall',
-  small: 'small',
-  medium: 'medium',
-  large: 'large',
-  extraLarge: 'extraLarge',
-  extraExtraLarge: 'extraExtraLarge'
-}
-export const ATOM_ICON_RENDERS = {
-  eager: 'eager',
-  lazy: 'lazy'
-}
-
-export default function AtomIcon({
+const AtomIcon = ({
   children,
   color = ATOM_ICON_COLORS.currentColor,
   size = ATOM_ICON_SIZES.small,
   render = ATOM_ICON_RENDERS.eager,
   title
-}) {
+}) => {
   const className = cx(
     BASE_CLASS,
     `${BASE_CLASS}--${size}`,
@@ -75,3 +58,7 @@ AtomIcon.propTypes = {
    */
   title: PropTypes.string
 }
+
+export default AtomIcon
+
+export {BASE_CLASS, ATOM_ICON_COLORS, ATOM_ICON_SIZES, ATOM_ICON_RENDERS}
