+++ components/atom/icon/src/settings.js
@@ -1,0 +1,27 @@
+export const BASE_CLASS = 'sui-AtomIcon'
+
+export const ATOM_ICON_COLORS = {
+  accent: 'accent',
+  alert: 'alert',
+  currentColor: 'currentColor',
+  error: 'error',
+  primary: 'primary',
+  success: 'success',
+  gray: 'gray'
+}
+
+export const ATOM_ICON_SIZES = {
+  extraSmall: 'extraSmall',
+  small: 'small',
+  medium: 'medium',
+  large: 'large',
+  extraLarge: 'extraLarge',
+  extraExtraLarge: 'extraExtraLarge'
+}
+
+export const ATOM_ICON_RENDERS = {
+  eager: 'eager',
+  lazy: 'lazy'
+}
+
+export const IMG_ROLE = 'img'
